+++ solvers/IncNavierStokesSolver/EquationSystems/IncNavierStokes.cpp
@@ -60,6 +60,20 @@ using namespace std;
 namespace Nektar
 {
 
+std::string IncNavierStokes::eqTypeLookupIds[6] = {
+    LibUtilities::SessionReader::RegisterEnumValue("EqType", "SteadyStokes",
+                                                   eSteadyStokes),
+    LibUtilities::SessionReader::RegisterEnumValue("EqType", "SteadyOseen",
+                                                   eSteadyOseen),
+    LibUtilities::SessionReader::RegisterEnumValue(
+        "EqType", "SteadyNavierStokes", eSteadyNavierStokes),
+    LibUtilities::SessionReader::RegisterEnumValue(
+        "EqType", "SteadyLinearisedNS", eSteadyLinearisedNS),
+    LibUtilities::SessionReader::RegisterEnumValue(
+        "EqType", "UnsteadyNavierStokes", eUnsteadyNavierStokes),
+    LibUtilities::SessionReader::RegisterEnumValue("EqType", "UnsteadyStokes",
+                                                   eUnsteadyStokes)};
+
 /**
  * Constructor. Creates ...
  *
@@ -112,22 +126,6 @@ void IncNavierStokes::v_InitObject(bool DeclareField)
         }
     }
     ASSERTL0(i != eEquationTypeSize, "EQTYPE not found in SOLVERINFO section");
-
-    // This probably should to into specific implementations
-    // Equation specific Setups
-    switch (m_equationType)
-    {
-        case eSteadyStokes:
-        case eSteadyOseen:
-        case eSteadyNavierStokes:
-        case eSteadyLinearisedNS:
-        case eUnsteadyNavierStokes:
-        case eUnsteadyStokes:
-            break;
-        case eNoEquationType:
-        default:
-            ASSERTL0(false, "Unknown or undefined equation type");
-    }
 
     m_session->LoadParameter("Kinvis", m_kinvis);